commit 8

+++ BE-LosQuebrachosApp/Migrations/20221112150649_Initial.cs
@@ -33,7 +33,7 @@ namespace BE_LosQuebrachosApp.Migrations
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Nombre = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     Apellido = table.Column<string>(type: "nvarchar(max)", nullable: false),
-                    Cuit = table.Column<int>(type: "int", nullable: false)
+                    Cuit = table.Column<long>(type: "bigint", nullable: false)
                 },
                 constraints: table =>
                 {
@@ -48,7 +48,7 @@ namespace BE_LosQuebrachosApp.Migrations
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Nombre = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     Apellido = table.Column<string>(type: "nvarchar(max)", nullable: false),
-                    Cuit = table.Column<int>(type: "int", nullable: false),
+                    Cuit = table.Column<long>(type: "bigint", nullable: false),
                     TransporteId = table.Column<int>(type: "int", nullable: false)
                 },
                 constraints: table =>
